requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pandas
numpy
beautifulsoup4
requests
openpyxl
scikit-learn
torch